ETH Transaction Utils Refactors (#4827)
* Move validateTransactionForSigning into validation.ts

* Move detect_transaction_type to it's own file

* Move format_transaction to it's own file

* Move prepare_transaction_for_signing to it's own file

* Move transaction_builder to it's own file

* Remove eth_tx.ts

* Move format_transaction to utils/

* Move format_transaction to utils/

* Move format_transaction to utils

* Update import of tx utils

* Solve merge conflicts, eslint corrections
