+++ packages/web3-eth/src/index.ts
@@ -2,9 +2,11 @@ import { Web3Eth } from './web3_eth';
 
 export * from './web3_eth';
 export * from './types';
-export * from './format_transaction';
 export * from './validation';
 export * from './rpc_method_wrappers';
-export * from './eth_tx';
+export * from './utils/format_transaction';
+export * from './utils/prepare_transaction_for_signing';
+export { detectTransactionType } from './utils/detect_transaction_type';
+export { transactionBuilder } from './utils/transaction_builder';
 
 export default Web3Eth;
