+++ packages/web3-eth/test/unit/validate_transaction_for_signing.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable jest/no-conditional-expect */
 
 import { InvalidTransactionObjectError } from '../../src/errors';
-import { validateTransactionForSigning } from '../../src/eth_tx';
+import { validateTransactionForSigning } from '../../src/validation';
 import {
 	invalidNonceOrChainIdData,
 	invalidTransactionObject,
